+++ src/utils/Math.cpp
@@ -10,17 +10,17 @@ namespace math{
 // Vector3
 //===============================
 
-    double& Vector3::x() { return data[0]; }
-    const double& Vector3::x() const { return data[0]; }
+    float& Vector3::x() { return data[0]; }
+    const float& Vector3::x() const { return data[0]; }
     
-    double& Vector3::y() { return data[1]; }
-    const double& Vector3::y() const { return data[1]; }
+    float& Vector3::y() { return data[1]; }
+    const float& Vector3::y() const { return data[1]; }
     
-    double& Vector3::z() { return data[2]; }
-    const double& Vector3::z() const { return data[2]; }
+    float& Vector3::z() { return data[2]; }
+    const float& Vector3::z() const { return data[2]; }
 
-    double& Vector3::operator[](std::size_t idx) { return data[idx]; }
-    const double& Vector3::operator[](std::size_t idx) const { return data[idx]; }
+    float& Vector3::operator[](std::size_t idx) { return data[idx]; }
+    const float& Vector3::operator[](std::size_t idx) const { return data[idx]; }
 
     Vector3 Vector3::operator+(const Vector3& other) const {
         Vector3 result{};
